"Solitario Game". Final Project 1st year UCM.
I developed a game inspired by the globally renowned card game, Solitaire, utilizing skills acquired during my first year of studies in Computer Engineering. This project was chosen due its requirements of the use of logical techniques to ensure functionality. The game was developed in C++ and served as my final project for the first year at Universidad Complutense de Madrid. It was awarded a grade of 9.5 out of 10.